LayoutTests/fast/page-color-sampling/maintain-sampled-color-when-scrolled-to-top-expected.txt

PASS colorsAfterScrollingDown.top is "rgb(255, 99, 71)"
PASS colorsAfterScrollingBackUp.top is "rgb(255, 99, 71)"
PASS successfullyParsed is true

TEST COMPLETE
relative
